+++ NMEAShared.h
@@ -1,13 +1,15 @@
 #ifndef NMEAShared_h
 #define NMEAShared_h
 
-#include "Types.h"
+#include "types.h"
+#include <cstring>
 
-int calculateChecksum(char *message, int length);
+
+int calculateChecksum(char *message, size_t length);
 
 double degreesFromCoordinateString(const char *string, char direction);
 
-void splitMessageIntoFragments(const char *message, int messageLength, char **fragments, int *fragmentCount);
+void splitMessageIntoFragments(const char *message, size_t messageLength, char **fragments, int *fragmentCount);
 
 Time timeFromString(const char *timeString);
